81613: Do not delete unmerged handoff SDK
Fixes:
https://code.google.com/p/android/issues/detail?id=81613

(cherry picked from commit bf69d121480a0ceb96515958efaf46b0293f8805)

+++ android/src/com/android/tools/idea/welcome/InstallComponentsPath.java
@@ -346,6 +346,7 @@ public class InstallComponentsPath extends DynamicWizardPath implements LongRunn
   private static class MergeOperation extends InstallOperation<File, File> {
     private final File myRepo;
     private final InstallContext myContext;
+    private boolean isRepoNoLongerNeeded = false;
 
     public MergeOperation(File repo, InstallContext context, double progressRatio) {
       super(context, progressRatio);
@@ -361,8 +362,9 @@ public class InstallComponentsPath extends DynamicWizardPath implements LongRunn
         FileUtil.ensureExists(destination);
         if (!FileUtil.filesEqual(destination.getCanonicalFile(), myRepo.getCanonicalFile())) {
           SdkMerger.mergeSdks(myRepo, destination, indicator);
+          isRepoNoLongerNeeded = true;
         }
-        myContext.print(String.format("Android SDK was installed to %s", destination), ConsoleViewContentType.SYSTEM_OUTPUT);
+        myContext.print(String.format("Android SDK was installed to %s\n", destination), ConsoleViewContentType.SYSTEM_OUTPUT);
         return destination;
       }
       catch (IOException e) {
@@ -375,7 +377,7 @@ public class InstallComponentsPath extends DynamicWizardPath implements LongRunn
 
     @Override
     public void cleanup(@NotNull File result) {
-      if (myRepo.exists()) {
+      if (isRepoNoLongerNeeded && myRepo.exists()) {
         FileUtil.delete(myRepo);
       }
     }
